Add component folder, format code
