--- frontend/src/Dummy.js
@@ -1,18 +1,0 @@
-import 'react-bootstrap';
-import { Card } from 'react-bootstrap';
-import { Jumbotron } from 'react-bootstrap';
-import { Button } from 'react-bootstrap';
-import { Badge } from 'react-bootstrap';
-import styles from './styles.css';
-
-export default function DummyPage() {
-    return (
-        <>
-            <div>
-            </div>
-        </>
-
-    )
-
-
-}
