+++ frontend/src/components/Dummy.js
@@ -1,0 +1,9 @@
+import "react-bootstrap";
+
+export default function DummyPage() {
+    return (
+        <>
+            <div></div>
+        </>
+    );
+}
