+++ frontend/src/components/HomePage.js
@@ -1,5 +1,6 @@
 import Footer from './Footer';
 import Body from './Body';
+
 export default function HomePage() {
     return (
         <>
